# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    OmniProject API — the stable, broker-agnostic consumer (northbound) API. It
    surfaces projects, issues, programmes, reports and admin operations above the
    swappable broker seam, so it stays the same regardless of which broker reaches your backends. The southbound broker contract a broker must
    implement is published separately at GET /api/contract.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      description: Fetches normalized project list via the broker
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (brokered to the backend; manager+)
      description: >-
        Creates a project in the backend system of record via the broker. Only
        available when the backend can store projects
        (capabilities.entities.project.store). Setting programmeId here is how a
        project joins a programme.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid input
        "403":
          description: Not permitted, or the backend can't store projects
...
  /settings:
    get:
      operationId: getSettings
      tags: [proxy]
      summary: Get integration settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [proxy]
      summary: Update integration settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        identifier:
          type: string
        description:
          type: ["string", "null"]
        source:
          type: string
          description: plane or openproject
        programmeId:
          type: ["string", "null"]
          description: >-
            Optional programme this project belongs to (owned by the backend). A
            project with no programmeId is standalone; OmniProject derives the
            programme grouping from this field, so a programme exists only when at
            least one project references it.
        programmeName:
          type: ["string", "null"]
        omniInstanceId:
          type: ["string", "null"]
          description: >-
            OmniProject's backend-independent correlation GUID for this project
            (minted at creation). Drives programme membership, the closed-project
            registry and the on-close disposition flow.
        status:
          type: ["string", "null"]
          description: Lifecycle status (backend-native); live reads exclude closed/archived.
        issueCount:
          type: integer
        completedCount:
          type: integer
        memberCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - identifier
        - source
        - issueCount
        - completedCount
        - memberCount
        - updatedAt
    ProjectInput:
      type: object
      description: Fields to create a project (brokered to the backend).
      properties:
        name:
          type: string
        identifier:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        programmeId:
          type: ["string", "null"]
          description: Optional programme to join on creation.
        status:
          type: ["string", "null"]
          description: >
            Optional lifecycle status (backend-native string). Conventional buckets are
            live/active, closed and archived; live reads exclude closed/archived. Set/clear to
            move a project between live and closed.
      required:
        - name
...
    FieldSupport:
      type: object
      description: Whether a field/entity can be surfaced (read) and stored (written).
      properties:
        surface:
          type: boolean
        store:
          type: boolean
      required:
        - surface
        - store
...
    Settings:
      type: object
      properties:
        brokerUrl:
          type: ["string", "null"]
          description: The active broker's webhook/endpoint URL.
        aiProvider:
          type: string
          enum: [none, openai, ollama, anthropic, openrouter]
        sttProvider:
          type: string
          description: >-
            Speech-to-text engine. "browser" runs on-device (no audio egress);
            "whisper" is an AI-assisted, off-device transcription provider
            (governance-gated). "none" disables dictation server-side.
          enum: [none, browser, whisper]
        aiModel:
          type: ["string", "null"]
        backendSource:
          type: string
          description: >-
            Free-form backend routing hint passed to the broker (e.g. "all",
            "jira", "azure-devops", "servicenow", "plane", "openproject"). "all"
            means no filter — whatever the broker is wired to.
        reportingCurrency:
          type: ["string", "null"]
          description: >-
            ISO 4217 code the consolidated financial reports default to (e.g. "GBP").
            Empty/absent ⇒ the FX table's base currency is used. Display-only — amounts
            are converted at view time via the broker FX rates; nothing is re-stored.
        fxRatePolicy:
          type: string
          description: >-
            Which FX rate consolidated reports convert at: "spot" (today's live rate),
            "periodClose" (the rate as of fxRateAsOfDate, e.g. the period the books
            closed at) or "budgetRate" (the rate as of fxRateAsOfDate treated as the
            rate the budget was set at, so variance isn't polluted by FX drift). A
            broker that can't serve a historical rate degrades to its live snapshot.
          enum: [spot, periodClose, budgetRate]
        fxRateAsOfDate:
          type: ["string", "null"]
          description: >-
            ISO 8601 date the "as of" rate is read for when fxRatePolicy isn't "spot".
            Ignored (falls back to spot) when null.
        oidcIssuerUrl:
          type: ["string", "null"]
        deploymentProfile:
          type: string
          description: >-
            Deployment context chosen in the setup wizard, which relaxes enterprise
            couplings by choice (e.g. a charity/self-hosted instance on a plain-HTTP LAN).
            Optional; absent until an admin selects one. The infra-level DEPLOYMENT_PROFILE
            env var takes precedence on a fresh boot (see docs/REVERSE-PROXY.md).
          enum: [enterprise, business, nonprofit, self-hosted, demo]
        fieldOverrides:
          $ref: "#/components/schemas/FieldMapOverride"
      required:
        - aiProvider
        - backendSource
    FieldMapOverride:
      type: object
      description: >-
        Admin translation-layer overrides for the capability map. Each entry
        REPLACES that field's/entity's surface+store, correcting a mis-mapping.
      properties:
        fields:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FieldSupport"
        entities:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FieldSupport"
    SettingsUpdate:
      type: object
      properties:
        brokerUrl:
          type: ["string", "null"]
        aiProvider:
          type: string
          enum: [none, openai, ollama, anthropic, openrouter]
        sttProvider:
          type: string
          description: >-
            Speech-to-text engine ("none" | "browser" on-device | "whisper" AI-assisted).
          enum: [none, browser, whisper]
        aiModel:
          type: ["string", "null"]
        backendSource:
          type: string
          description: >-
            Free-form backend routing hint passed to the broker (see Settings.backendSource).
        reportingCurrency:
          type: ["string", "null"]
          description: >-
            Default ISO 4217 reporting currency for consolidated financial reports
            (see Settings.reportingCurrency). Empty/null clears it (falls back to the FX base).
        fxRatePolicy:
          type: string
          description: Set the FX as-of-date policy (see Settings.fxRatePolicy).
          enum: [spot, periodClose, budgetRate]
        fxRateAsOfDate:
          type: ["string", "null"]
          description: Set the "as of" date used when fxRatePolicy isn't "spot" (see Settings.fxRateAsOfDate).
        oidcIssuerUrl:
          type: ["string", "null"]
        deploymentProfile:
          type: string
          description: >-
            Set the deployment profile (admin). Persisted; the infra-level
            DEPLOYMENT_PROFILE env var still wins on a fresh boot (see docs/REVERSE-PROXY.md).
          enum: [enterprise, business, nonprofit, self-hosted, demo]
        fieldOverrides:
          $ref: "#/components/schemas/FieldMapOverride"
...